requirements.txt
Flask==3.0.0
flask-cors==4.0.0
GitPython==3.1.40
requests==2.31.0